+++ VSDocumentReopen/VSDocumentReopen.Test/Infrastructure/JsonHistoryRepositoryFactoryTest.cs
@@ -1,42 +1,35 @@
-﻿using Moq;
-using System.ComponentModel;
+﻿using System.ComponentModel;
 using VSDocumentReopen.Domain;
 using VSDocumentReopen.Infrastructure;
 using VSDocumentReopen.Infrastructure.Repositories;
+using VSDocumentReopen.Test.AssemblyFictures;
 using Xunit;
 
 namespace VSDocumentReopen.Test.Infrastructure
 {
-	public class JsonHistoryRepositoryFactoryTest
+	public class JsonHistoryRepositoryFactoryTest : IAssemblyFixture<ConfigContext>
 	{
-		private readonly Mock<IConfiguration> _configurationMock;
 		private readonly JsonHistoryRepositoryFactory _jsonHistoryRepositoryFactory;
-		private readonly Mock<ConfigurationManager> _configurationManagerMock;
+		private readonly ConfigContext _configContext;
 
-		public JsonHistoryRepositoryFactoryTest()
+		public JsonHistoryRepositoryFactoryTest(ConfigContext configContext)
 		{
-			_configurationMock = new Mock<IConfiguration>();
-			_configurationManagerMock = new Mock<ConfigurationManager>();
-			_configurationManagerMock.SetupGet(g => g.Config).Returns(_configurationMock.Object);
-
-			ConfigurationManager.Current = _configurationManagerMock.Object;
-
 			_jsonHistoryRepositoryFactory = new JsonHistoryRepositoryFactory(null);
+			_configContext = configContext;
 		}
 
 		[Fact]
 		public void ItShould_Handle_Null()
 		{
-			Assert.Throws<InvalidEnumArgumentException>(() =>
-				_jsonHistoryRepositoryFactory.CreateHistoryRepository(null));
+			Assert.Throws<InvalidEnumArgumentException>(() => _jsonHistoryRepositoryFactory.CreateHistoryRepository(null));
 		}
 
 		[Fact]
 		public void ItShould_Handle_ValidObject()
 		{
-			_configurationMock.SetupGet(g => g.VSTempFolderName).Returns("");
-			_configurationMock.SetupGet(g => g.PackageWorkingDirName).Returns("");
-			_configurationMock.SetupGet(g => g.HistoryFileName).Returns("");
+			_configContext.Configuration.SetupGet(g => g.VSTempFolderName).Returns("VSTempFolderName");
+			_configContext.Configuration.SetupGet(g => g.PackageWorkingDirName).Returns("PackageWorkingDirName");
+			_configContext.Configuration.SetupGet(g => g.HistoryFileName).Returns("HistoryFileName");
 
 			var ret = _jsonHistoryRepositoryFactory.CreateHistoryRepository(new SolutionInfo("c:\\", "test"));