+++ VSDocumentReopen/VSDocumentReopen/Infrastructure/Version/VsVersionContext.cs
@@ -1,0 +1,9 @@
+﻿namespace VSDocumentReopen.Infrastructure.Version
+{
+	public abstract class VsVersionContext
+	{
+		public static VsVersionContext Current { get; set; }
+
+		public abstract IVsVersionProvider VersionProvider { get; }
+	}
+}