+++ VSDocumentReopen/VSDocumentReopen/VS/ToolWindows/ClosedDocumentsHistoryControl.xaml.cs
@@ -1,9 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using System.Windows.Controls;
 using System.Windows.Media.Imaging;
 using VSDocumentReopen.Domain.Documents;
+using VSDocumentReopen.Domain.HistoryControl;
+using VSDocumentReopen.Infrastructure;
 using VSDocumentReopen.Infrastructure.Document.Tracking;
 using VSDocumentReopen.Infrastructure.FileIcons;
 using VSDocumentReopen.Infrastructure.HistoryCommands;
@@ -15,10 +18,9 @@ namespace VSDocumentReopen.VS.ToolWindows
 	/// <summary>
 	/// Interaction logic for ClosedDocumentsHistoryControl.
 	/// </summary>
-	public partial class ClosedDocumentsHistoryControl : UserControl
+	public partial class ClosedDocumentsHistoryControl : UserControl, IDisposable
 	{
-		private static readonly Dictionary<string, BitmapSource> _fileTypeImages
-			= new Dictionary<string, BitmapSource>();
+		private static readonly Dictionary<string, BitmapSource> _fileTypeImages = new Dictionary<string, BitmapSource>();
 
 		private readonly Func<IClosedDocument, bool> GetFullHistory = _ => true;
 
@@ -28,7 +30,7 @@ namespace VSDocumentReopen.VS.ToolWindows
 		private readonly IHistoryCommandFactory _removeSomeDocumentsCommandFactory;
 		private readonly IHistoryCommand _clearHistoryCommand;
 		private readonly IFileExtensionIconResolver _fileExtensionIconResolver;
-
+		private readonly IHistoryToolWindowRepositoryFactory _historyToolWindowRepositoryFactory;
 		private GridViewColumnHeader listViewSortCol = null;
 		private SortAdorner listViewSortAdorner = null;
 
@@ -42,7 +44,8 @@ namespace VSDocumentReopen.VS.ToolWindows
 			IHistoryCommandFactory reopenSomeDocumentsCommandFactory,
 			IHistoryCommandFactory removeSomeDocumentsCommandFactory,
 			IHistoryCommand clearHistoryCommand,
-			IFileExtensionIconResolver fileExtensionIconResolver)
+			IFileExtensionIconResolver fileExtensionIconResolver,
+			IHistoryToolWindowRepositoryFactory historyToolWindowRepositoryFactory)
 		{
 			InitializeComponent();
 
@@ -52,26 +55,28 @@ namespace VSDocumentReopen.VS.ToolWindows
 			_removeSomeDocumentsCommandFactory = removeSomeDocumentsCommandFactory;
 			_clearHistoryCommand = clearHistoryCommand;
 			_fileExtensionIconResolver = fileExtensionIconResolver;
+			_historyToolWindowRepositoryFactory = historyToolWindowRepositoryFactory;
 
 			var openState = new ButtonDisabledState(_openSelected,
-				new Image() {Source = WpfImageSourceConverter.CreateBitmapSource(VSDocumentReopen.Resources.OpenFile_16x)},
-				new Image() {Source = WpfImageSourceConverter.CreateBitmapSource(VSDocumentReopen.Resources.OpenFile_16x_Gray)});
+				new Image() { Source = WpfImageSourceConverter.CreateBitmapSource(VSDocumentReopen.Resources.OpenFile_16x) },
+				new Image() { Source = WpfImageSourceConverter.CreateBitmapSource(VSDocumentReopen.Resources.OpenFile_16x_Gray) });
 			openState.Disable();
 
 			var removeState = new ButtonDisabledState(_removeSelected,
-				new Image() {Source = WpfImageSourceConverter.CreateBitmapSource(VSDocumentReopen.Resources.RemoveGuide_16x)},
-				new Image() {Source = WpfImageSourceConverter.CreateBitmapSource(VSDocumentReopen.Resources.RemoveGuide_16x_Gray)});
+				new Image() { Source = WpfImageSourceConverter.CreateBitmapSource(VSDocumentReopen.Resources.RemoveGuide_16x) },
+				new Image() { Source = WpfImageSourceConverter.CreateBitmapSource(VSDocumentReopen.Resources.RemoveGuide_16x_Gray) });
 			removeState.Disable();
 
 			var clearState = new ButtonDisabledState(_clearAll,
-				new Image() {Source = WpfImageSourceConverter.CreateBitmapSource(VSDocumentReopen.Resources.ClearWindowContent_16x)},
-				new Image() {Source = WpfImageSourceConverter.CreateBitmapSource(VSDocumentReopen.Resources.ClearWindowContent_16x_Gray)});
+				new Image() { Source = WpfImageSourceConverter.CreateBitmapSource(VSDocumentReopen.Resources.ClearWindowContent_16x) },
+				new Image() { Source = WpfImageSourceConverter.CreateBitmapSource(VSDocumentReopen.Resources.ClearWindowContent_16x_Gray) });
 			clearState.Disable();
 
 			_documentHistoryQueries.HistoryChanged += DocumentHistoryChanged;
 			UpdateHistoryView(GetFullHistory);
 
 			_listView.Focus();
+			LoadSettings();
 		}
 
 		private void DocumentHistoryChanged(object sender, EventArgs e)
@@ -92,6 +97,26 @@ namespace VSDocumentReopen.VS.ToolWindows
 			}
 
 			return _fileTypeImages[extension];
+		}
+
+		private void LoadSettings()
+		{
+			var settingsRepository = _historyToolWindowRepositoryFactory.Create();
+			var settings = settingsRepository?.GetSettings();
+
+			_search.HistoryList = settings?.SearchHistory?.ToList() ?? new List<string>();
+		}
+
+		public void Dispose()
+		{
+			var settingsRepository = _historyToolWindowRepositoryFactory.Create();
+
+			var settings = new HistoryControlData()
+			{
+				SearchHistory = _search.HistoryList.Take(10),
+			};
+
+			settingsRepository?.SaveSettings(settings);
 		}
 	}
 }